Major refactor. Cut out some of the needlessly complex code (throw/catch being a good example). Tried to make Token (and now Tokens) more object orientated and cut down on interim calls by hiding them behind the Token(s) interface.
Moved the algorithms to their own file because it's easier to read
the main Handler class, and making them lambdas means the can be
overriden easier or added too, and hopefully specced better.

The Repeater didn't need to be its own object, it was only running
in the the Token class so made it an included module.

Updated the spec for Token.

*Still* doesn't handle the initial problem with "at" nor has this
fixed the parsing of "the 3rd sunday of the month", but at least
it should be easier to find out why.