Change the dependency from chronic to gitlab-chronic
As the regular chronic gem has a maintainer that isn't able to spend
time on it, Gitlab forked it and continued the maintenance.

Tickles test suite passes after moving to the gitlab-chronic gem and
with that development should be possible to continue without having to
wait on the missing features in regular chronic.